Introduce IBaseRefactoring<TModel> and split ReorderParameterRefactoring using it
