+++ RubberduckTests/Commands/RefactorCommands/ReorderParametersCommandTests.cs
@@ -170,7 +170,8 @@ End Property";
             uiDispatcherMock
                 .Setup(m => m.Invoke(It.IsAny<Action>()))
                 .Callback((Action action) => action.Invoke());
-            var refactoring = new ReorderParametersRefactoring(state, factory, rewritingManager, selectionService, selectedDeclarationProvider, uiDispatcherMock.Object);
+            var baseRefactoring = new ReorderParameterBaseRefactoring(state, rewritingManager);
+            var refactoring = new ReorderParametersRefactoring(baseRefactoring, state, factory, rewritingManager, selectionService, selectedDeclarationProvider, uiDispatcherMock.Object);
             var notifier = new ReorderParametersFailedNotifier(msgBox);
             return new RefactorReorderParametersCommand(refactoring, notifier, state, selectionService, selectedDeclarationProvider);
         }
